+++ packages/core/examples/customNode/cusomHexagon.mjs
@@ -1,3 +1,5 @@
+// import { h } from "preact";
+
 class CustomHexagonModel extends PolygonNodeModel {
   setAttributes() {
     const width = 100;
@@ -14,6 +16,7 @@ class CustomHexagonModel extends PolygonNodeModel {
       [x - 0.5 * width, y],
     ];
     this.points = pointList;
+    this.text.editable = false;
   }
   getConnectedSourceRules() {
     const rules = super.getConnectedSourceRules();
@@ -38,8 +41,21 @@ class CustomHexagonModel extends PolygonNodeModel {
   }
 }
 
+class CusomHexagonView extends PolygonNode {
+  getAnchorShape(anchorData) {
+    const { x, y } = anchorData;
+    return h('rect', {
+      x: x - 5,
+      y: y - 5,
+      width: 10,
+      height: 10,
+      className: 'custom-anchor'
+    })
+  }
+}
+
 export default {
   type: 'hexagonNode',
   model: CustomHexagonModel,
-  view: PolygonNode
+  view: CusomHexagonView
 }