+++ packages/extension/src/NodeResize/Node/DiamondResize.tsx
@@ -18,6 +18,14 @@ class DiamondResizeModel extends DiamondNodeModel {
   minHeight = 30;
   maxWidth = 2000;
   maxHeight = 2000;
+  constructor(data, graphModel) {
+    super(data, graphModel);
+    const { nodeSize } = this.properties;
+    if (nodeSize) {
+      this.rx = nodeSize.rx;
+      this.ry = nodeSize.ry;
+    }
+  }
   getOutlineStyle() {
     const style = super.getOutlineStyle();
     style.stroke = 'none';
@@ -40,14 +48,6 @@ class DiamondResizeModel extends DiamondNodeModel {
       fill: '#FFFFFF',
       stroke: '#000000',
     };
-  }
-  setAttributes() {
-    // @ts-ignore
-    const { nodeSize } = this.properties;
-    if (nodeSize) {
-      this.rx = nodeSize.rx;
-      this.ry = nodeSize.ry;
-    }
   }
 }
 class DiamondResizeView extends DiamondNode {