+++ site/mvp/src/components/node/arrow/RightArrow.js
@@ -4,8 +4,8 @@ import RectNode from '../basic/RectNode'
 // 右箭头
 
 class RightArrowModel extends RectNode.model {
-  constructor(data, graphData) {
-    super(data, graphData)
+  initNodeData(data) {
+    super.initNodeData(data)
     this.width = 80
     this.height = 50
   }